# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UVA Production Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rise/bank:
    get:
      operationId: getRiseBank
      tags: [rise]
      summary: Get the Rise item bank for a level (no answer keys)
      parameters:
        - name: level
          in: query
          required: true
          schema:
            type: string
            enum: [elementary, secondary, higher]
      responses:
        "200":
          description: Item bank
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiseBank"
  /rise/answer:
    post:
      operationId: answerRiseItem
      tags: [rise]
      summary: Grade a single Rise answer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiseAnswerInput"
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiseAnswerResult"
  /rise/sessions:
    get:
      operationId: getRiseSessions
      tags: [rise]
      summary: List the learner's past Rise runs
      responses:
        "200":
          description: Sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RiseSession"
    post:
      operationId: saveRiseSession
      tags: [rise]
      summary: Persist a completed Rise run for the signed-in learner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiseSessionInput"
      responses:
        "201":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiseSession"

  # ── Meridian (provider operations) ───────────────────────────
...
  /portal/messages:
    post:
      operationId: sendPortalMessage
      tags: [portal]
      summary: Send a message to the engagement team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortalMessageInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitResult"

  # ── Admin ────────────────────────────────────────────────────
...
components:
  schemas:
...
    SubmitResult:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        id:
          type: integer

    # ── Contact & newsletter ──────────────────────────────────────
...
    RiseBankItem:
      type: object
      required: [id, difficulty, skill, passage, question, options]
      properties:
        id:
          type: string
        difficulty:
          type: integer
        skill:
          type: string
        passage:
          type: string
        question:
          type: string
        options:
          type: array
          items:
            type: string
    RiseBank:
      type: object
      required: [level, items]
      properties:
        level:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/RiseBankItem"
    RiseAnswerInput:
      type: object
      required: [itemId, optionIndex]
      properties:
        itemId:
          type: string
        optionIndex:
          type: integer
    RiseAnswerResult:
      type: object
      required: [correct, correctIndex]
      properties:
        correct:
          type: boolean
        correctIndex:
          type: integer
        hint:
          type: ["string", "null"]
    RisePathStep:
      type: object
      required: [itemId, difficulty, correct]
      properties:
        itemId:
          type: string
        difficulty:
          type: integer
        correct:
          type: boolean
    RiseSessionInput:
      type: object
      required: [level, itemsAttempted, correctCount, masteryEstimate, path]
      properties:
        level:
          type: string
          enum: [elementary, secondary, higher]
        itemsAttempted:
          type: integer
        correctCount:
          type: integer
        masteryEstimate:
          type: integer
        finalRung:
          type: string
        path:
          type: array
          items:
            $ref: "#/components/schemas/RisePathStep"
    RiseSession:
      type: object
      required:
        [id, level, itemsAttempted, correctCount, masteryEstimate, path, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        level:
          type: string
        itemsAttempted:
          type: integer
        correctCount:
          type: integer
        masteryEstimate:
          type: integer
        finalRung:
          type: ["string", "null"]
        path:
          type: array
          items:
            $ref: "#/components/schemas/RisePathStep"
        createdAt:
          type: string
          format: date-time

    # ── Meridian (provider operations) ────────────────────────────
...
    PortalMessageInput:
      type: object
      required: [subject, message]
      properties:
        subject:
          type: string
        message:
          type: string

    # ── Admin ─────────────────────────────────────────────────────
...