# lib/api-zod/src/generated/types/riseAnswerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * UVA Production Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface RiseAnswerInput {
  itemId: string;
  optionIndex: number;
}