# lib/api-zod/src/generated/types/riseSessionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * UVA Production Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { RisePathStep } from './risePathStep';
import type { RiseSessionInputLevel } from './riseSessionInputLevel';

export interface RiseSessionInput {
  level: RiseSessionInputLevel;
  itemsAttempted: number;
  correctCount: number;
  masteryEstimate: number;
  finalRung?: string;
  path: RisePathStep[];
}